I want to create a number of 'gift' pages, one for each gift in @/packages\ui\src\components\GiftsSection.tsx
the urls should be something like /gift/[gift-slug]

For this,

1. extract the gift data from gift section and also create a separate type. this should also support the properties: authors [array: name, link], slug - we keep link, but we render the internal link using the slug in GiftsSection
2. Create a new file in pages that loads in the gift slug dynamically and then renders the gift
base the content originally on what we have in the gift section. But this first has the handwritten text: 'Our gift for you' - and after the description: made by [authors]
after that CTA [gift icon | get it now] (base this on the learn more button from the gift section)
this should take up the whole page. then have a hovering button 'learn more' - base this on the scroll indicator we have in @/packages\ui\src\pages\index.tsx
this goes to a handwritten heading 'What we've made for you' with the anchor '#details'